openvpn_status: rename collector and files
