+++ config/go.d.conf
@@ -77,3 +77,4 @@ modules:
 #  wmi: yes
 #  x509check: yes
 #  zookeeper: yes
+#  openvpn_status: yes
