+++ modules/init.go
@@ -36,7 +36,7 @@ import (
 	_ "github.com/netdata/go.d.plugin/modules/nginx"
 	_ "github.com/netdata/go.d.plugin/modules/nginxvts"
 	_ "github.com/netdata/go.d.plugin/modules/openvpn"
-	_ "github.com/netdata/go.d.plugin/modules/ovpn_status_log"
+	_ "github.com/netdata/go.d.plugin/modules/openvpn_status"
 	_ "github.com/netdata/go.d.plugin/modules/phpdaemon"
 	_ "github.com/netdata/go.d.plugin/modules/phpfpm"
 	_ "github.com/netdata/go.d.plugin/modules/pihole"
